drivers/can: Fix compiler error
Cast size_t to unsigned long int.

+++ drivers/can/can.c
@@ -507,7 +507,7 @@ static ssize_t can_read(FAR struct file *filep, FAR char *buffer,
   FAR struct can_dev_s     *dev = inode->i_private;
 #endif
 
-  caninfo("buflen: %lu\n", buflen);
+  caninfo("buflen: %lu\n", (unsigned long)buflen);
 
   /* The caller must provide enough memory to catch the smallest possible
    * message.  This is not a system error condition, but we won't permit
@@ -744,7 +744,7 @@ static ssize_t can_write(FAR struct file *filep, FAR const char *buffer,
   int                      msglen;
   int                      ret   = 0;
 
-  caninfo("buflen: %lu\n", buflen);
+  caninfo("buflen: %lu\n", (unsigned long)buflen);
 
   /* Interrupts must disabled throughout the following */
